% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{dv_playlist_as_onclick}
\alias{dv_playlist_as_onclick}
\title{Convert playlist to 'onclick' string}
\usage{
dv_playlist_as_onclick(playlist, video_id)
}
\arguments{
\item{playlist}{data.frame: a playlist as returned by \code{dv_video_playlist}}

\item{video_id}{string: the id of the HTML video element to attach the playlist to}
}
\value{
A string suitable for inclusion as an 'onclick' tag attribute
}
\description{
Convert playlist to 'onclick' string
}
\examples{
\dontrun{
  library(shiny)

  ## hand-crafted playlist for this example
  playlist <- data.frame(video_src = "xL7qEpOdtio",
                         start_time = c(5417, 7252, 6222, 7656, 7369),
                         duration = 8,
                         type = "youtube")
  shinyApp(
      ui = fluidPage(
          dv_video_js(youtube = TRUE),
          dv_video_player(id = "yt_player", style = "height: 480px; background-color: black;"),
          tags$button("Go", onclick = dv_playlist_as_onclick(playlist, "yt_player"))
      ),
      server = function(input, output) {},
  )
}

}
