% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{dv_shiny_video_sync}
\alias{dv_shiny_video_sync}
\title{Launch a Shiny app to help with video synchronisation}
\usage{
dv_shiny_video_sync(dvw, video_file = NULL, launch_browser = TRUE, ...)
}
\arguments{
\item{dvw}{string or datavolley: either the path to a dvw file (which will be read by \code{\link[datavolley]{read_dv}}) or a datavolley object as returned by that function}

\item{video_file}{string: optionally, the path to the video file. If not supplied (or \code{NULL}) the video file specified in the dvw file will be used}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny]{runApp}}'s \code{launch.browser} parameter)}

\item{...}{: extra parameters passed to \code{\link[datavolley]{read_dv}} (if \code{dvw} is a provided as a string)}
}
\description{
Launch a Shiny app to help with video synchronisation
}
\seealso{
\code{\link[datavolley]{read_dv}}
}
