% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{dv_video_js}
\alias{dv_video_js}
\title{Inject javascript for an HTML video player}
\usage{
dv_video_js(youtube = FALSE)
}
\arguments{
\item{youtube}{logical: set to \code{TRUE} to include the Youtube API javascript}
}
\value{
A tag list
}
\description{
Inject javascript for an HTML video player
}
\seealso{
\code{\link{dv_video_playlist}}
}
